+++ src/components/ui/message/message.js
@@ -1,7 +1,7 @@
-export function message(messageShow,messageType,messageTextOverview,messageTextDetails){
+export function message(messageShow,messageType,messageTextOverview,messageTextDetails,blockApp){
 
     if(messageTextDetails === '' || !messageTextDetails){
-        messageTextDetails = 'Keine Details verfügbar.'
+        messageTextDetails = window.languag.message_noDetailsAvailable
     }
 
     let currentMessageID = new Date().getTime().toString()
@@ -49,6 +49,9 @@ export function message(messageShow,messageType,messageTextOverview,messageTextD
         setTimeout(function(){
             currentMessage.classList.remove('message--hidden')
         },300)
+    }
+    if(blockApp){
+        document.body.style.pointerEvents = 'none'
     }
 }