+++ src/dlc/dlc.js
@@ -1,0 +1,10 @@
+import { xmasDlc } from "./xmas/xmas.import.js"
+import { platform_functionName_automatic, platform_bookingPlatformPreValue } from "./platforms/platforms.import.js"
+import { platforms, platformsContent } from "./platforms/platforms.dlc.js"
+import { filters, filtersContent } from "./filters/filters.dlc.js"
+import { filter_timesheetFilterPreValue } from "./filters/filters.import.js"
+
+export {platform_functionName_automatic, 
+        platformsContent, platforms, platform_bookingPlatformPreValue, filters, 
+        filtersContent, filter_timesheetFilterPreValue, xmasDlc
+}